JBEHAVE-1314 Remove redundant type casts

+++ jbehave-guice/src/test/java/org/jbehave/core/steps/guice/GuiceStepsFactoryBehaviour.java
@@ -55,7 +55,7 @@ public class GuiceStepsFactoryBehaviour {
         List<CandidateSteps> steps = factory.createCandidateSteps();
         // Then
         assertFooStepsFound(steps);
-        assertThat((int) ((FooStepsWithDependency) stepsInstance(steps.get(0))).integer, equalTo(42));
+        assertThat(((FooStepsWithDependency) stepsInstance(steps.get(0))).integer, equalTo(42));
     }
 
     private void assertFooStepsFound(List<CandidateSteps> steps) {
